Separate AutoForms implementation
